Add USSD

+++ models/Subscription.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Subscription struct {
-	SubID       int64     `gorm:"column:sub_id;primaryKey;autoIncrement"`
+	SubID       uint      `gorm:"column:sub_id;primaryKey;autoIncrement"`
 	Name        string    `gorm:"column:name;not null"`
 	Description string    `gorm:"column:description;not null"`
 	Amount      int64     `gorm:"column:amount;not null"`
